# docs/openapi.yaml
openapi: 3.0.0
info:
  title: API REST - Logística de Entregas
  version: 1.0.0
  description: API para gestionar pedidos, conductores y entregas de logística
  contact:
    name: Equipo LogisticaMorales
    email: api@logistica-morales.com
  license:
    name: MIT

servers:
  - url: https://api.logistica-morales.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Development

tags:
  - name: Auth
    description: Autenticación de usuarios
  - name: Pedidos
    description: Gestión de pedidos
  - name: Conductor
    description: Operaciones de conductor
  - name: Entregas
    description: Registro de entregas

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en login

  schemas:
    Usuario:
      type: object
      required:
        - id
        - nombre
        - email
        - rol
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        telefono:
          type: string
          pattern: '^\+?[0-9]{7,15}$'
        rol:
          type: string
          enum: [admin, conductor]
        estado:
          type: string
          enum: [activo, inactivo]
        fotoUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Cliente:
      type: object
      required:
        - id
        - nombre
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
        telefono:
          type: string
        razonSocial:
          type: string
        ruc:
          type: string
          pattern: '^\d{11}$'

    ItemPedido:
      type: object
      required:
        - descripcion
        - cantidad
        - precioUnitario
      properties:
        id:
          type: string
          format: uuid
        descripcion:
          type: string
          minLength: 1
        cantidad:
          type: number
          minimum: 0.01
          multipleOf: 0.01
        precioUnitario:
          type: number
          minimum: 0
          multipleOf: 0.01
        cantidadLevantada:
          type: number
          minimum: 0
          multipleOf: 0.01
        estado:
          type: string
          enum: [pendiente, parcial, completado]

    Pedido:
      type: object
      required:
        - clienteId
        - numero
        - direccion
        - montoTotal
        - items
      properties:
        id:
          type: string
          format: uuid
        numero:
          type: string
          pattern: '^PED-\d{4}-\d{3}$'
        clienteId:
          type: string
          format: uuid
        conductorId:
          type: string
          format: uuid
          nullable: true
        direccion:
          type: string
          minLength: 5
        referencia:
          type: string
        latitud:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitud:
          type: number
          format: double
          minimum: -180
          maximum: 180
        montoTotal:
          type: number
          minimum: 0
          multipleOf: 0.01
        montoLevantado:
          type: number
          minimum: 0
          multipleOf: 0.01
        porcentajeLevantado:
          type: number
          minimum: 0
          maximum: 100
          multipleOf: 0.01
        estado:
          type: string
          enum: [pendiente, asignado, en_ruta, parcial, completado, cancelado]
        levantadoTotal:
          type: boolean
        observaciones:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemPedido'
        fechaCreacion:
          type: string
          format: date-time
        fechaAsignacion:
          type: string
          format: date-time
          nullable: true
        fechaEntregaEstimada:
          type: string
          format: date
        fechaEntregaReal:
          type: string
          format: date-time
          nullable: true

    Entrega:
      type: object
      required:
        - pedidoId
        - cantidadLevantada
        - recibidoPor
      properties:
        id:
          type: string
          format: uuid
        pedidoId:
          type: string
          format: uuid
        itemPedidoId:
          type: string
          format: uuid
        conductorId:
          type: string
          format: uuid
        cantidadLevantada:
          type: number
          minimum: 0.01
          multipleOf: 0.01
        fotoUrl:
          type: string
          format: uri
        fotoFirmaUrl:
          type: string
          format: uri
          nullable: true
        observaciones:
          type: string
        estado:
          type: string
          enum: [planeada, en_ruta, completada, rechazada]
        recibidoPor:
          type: string
          minLength: 1
        dniRecibidor:
          type: string
          pattern: '^\d{8}$'
        latitud:
          type: number
          format: double
        longitud:
          type: number
          format: double
        fechaEntrega:
          type: string
          format: date-time

    ApiResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          nullable: true
        error:
          type: object
          properties:
            code:
              type: string
            details:
              oneOf:
                - type: string
                - type: object

    ErrorResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - error
          properties:
            error:
              type: object
              required:
                - code
                - details
              properties:
                code:
                  type: string
                details:
                  type: object

paths:
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Login exitoso"
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      usuario:
                        $ref: '#/components/schemas/Usuario'
                      expiresIn:
                        type: integer
                        example: 86400
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Validación fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /pedidos:
    post:
      tags:
        - Pedidos
      summary: Crear nuevo pedido
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - clienteId
                - numero
                - direccion
                - montoTotal
                - items
              properties:
                clienteId:
                  type: string
                  format: uuid
                numero:
                  type: string
                  pattern: '^PED-\d{4}-\d{3}$'
                direccion:
                  type: string
                referencia:
                  type: string
                latitud:
                  type: number
                longitud:
                  type: number
                montoTotal:
                  type: number
                  minimum: 0
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/ItemPedido'
      responses:
        '201':
          description: Pedido creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Pedido'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    get:
      tags:
        - Pedidos
      summary: Listar pedidos
      security:
        - BearerAuth: []
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [pendiente, asignado, en_ruta, parcial, completado, cancelado]
        - name: conductor
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      pedidos:
                        type: array
                        items:
                          $ref: '#/components/schemas/Pedido'
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer
                          pages:
                            type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

  /pedidos/{pedidoId}:
    get:
      tags:
        - Pedidos
      summary: Obtener pedido por ID
      security:
        - BearerAuth: []
      parameters:
        - name: pedidoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Pedido'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Pedidos
      summary: Actualizar estado del pedido
      security:
        - BearerAuth: []
      parameters:
        - name: pedidoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - estado
              properties:
                estado:
                  type: string
                  enum: [pendiente, asignado, en_ruta, parcial, completado, cancelado]
                observaciones:
                  type: string
      responses:
        '200':
          description: Pedido actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Pedido'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /pedidos/{pedidoId}/asignar:
    post:
      tags:
        - Pedidos
      summary: Asignar conductor
      security:
        - BearerAuth: []
      parameters:
        - name: pedidoId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - conductorId
              properties:
                conductorId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Conductor asignado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Pedido'
        '400':
          $ref: '#/components/responses/BadRequest'

  /conductor/mis-pedidos:
    get:
      tags:
        - Conductor
      summary: Obtener mis pedidos asignados
      security:
        - BearerAuth: []
      parameters:
        - name: estado
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Pedidos del conductor
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      pedidos:
                        type: array
                        items:
                          $ref: '#/components/schemas/Pedido'
                      resumen:
                        type: object

  /entregas:
    post:
      tags:
        - Entregas
      summary: Crear entrega con foto
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - pedidoId
                - cantidadLevantada
                - recibidoPor
                - foto
              properties:
                pedidoId:
                  type: string
                  format: uuid
                itemPedidoId:
                  type: string
                  format: uuid
                cantidadLevantada:
                  type: number
                observaciones:
                  type: string
                recibidoPor:
                  type: string
                dniRecibidor:
                  type: string
                foto:
                  type: string
                  format: binary
                fotoFirma:
                  type: string
                  format: binary
      responses:
        '201':
          description: Entrega registrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Entrega'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          description: Archivo muy grande

components:
  responses:
    BadRequest:
      description: Validación fallida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Unauthorized:
      description: Token inválido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Forbidden:
      description: Permiso denegado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'